pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "mdimg-transfer"
version = "0.1.0"
description = "A web application for processing Markdown files and optimizing images"
authors = ["Codeium <support@codeium.com>"]
readme = "README.md"
license = "MIT"
packages = [
    { include = "mdimg_transfer" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Multimedia :: Graphics",
    "Framework :: AsyncIO",
    "Framework :: Quart",
]

[tool.poetry.dependencies]
python = ">=3.9"
quart = ">=0.18.0"
hypercorn = ">=0.14.0"
boto3 = ">=1.26.0"
python-dotenv = ">=0.19.0"
requests = ">=2.31.0"
aiohttp = ">=3.8.0"
aiofiles = ">=23.1.0"
tqdm = ">=4.65.0"
urllib3 = ">=2.0.0"
Werkzeug = ">=2.0.0"
Pillow = ">=10.0.0"
python-magic = ">=0.4.27"
prometheus-client = ">=0.17.0"
SQLAlchemy = ">=2.0.0"
aiosqlite = ">=0.19.0"
psutil = ">=5.9.0"
html2text = ">=2024.2.0"

[tool.poetry.group.test.dependencies]
pytest = ">=8.0.0"
pytest-asyncio = ">=0.24.0"
pytest-cov = ">=6.0.0"
pytest-mock = ">=3.12.0"
pytest-benchmark = ">=4.0.0"
pytest-profiling = ">=1.7.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
